# api/openapi.yaml
openapi: 3.1.0
info:
  title: Alfred API
  version: "1.0.0"
  description: |
    Resource-oriented REST API for the Alfred agentic platform. This document is
    the single source of truth for the headless backend and the separate Svelte
    client. Generate a typed client with
    `openapi-typescript api/openapi.yaml -o src/lib/api/types.ts`, or fetch the
    served copy at `GET /api/v1/openapi.json` (unauthenticated). A narrative
    integration guide with code samples lives at `api/README.md`.

    ## Base URL and versioning

    Every path below is served under `/api/v1`. The unversioned ops endpoints
    `GET /healthz` (liveness) and `GET /readyz` (readiness) live outside this
    contract and outside `/api/v1`; see `deploy/README.md`.

    ## Authentication

    Send `Authorization: Bearer <api key>` on every request. Call
    `GET /auth-info` (unauthenticated) first: when it returns
    `{"required": false}` the server was started without an API key and accepts
    requests without the header. Browsers cannot set an Authorization header on
    `EventSource`, so for SSE you mint a short-lived single-use ticket with
    `POST /auth/sse-tickets` and pass it as the `?ticket=` query parameter.

    ## Errors

    Every error is RFC 9457 `application/problem+json` with at least `title` and
    `status`. `401` is a missing or wrong bearer token (or a spent/expired SSE
    ticket). `404` is an unknown run, config, or task. `409` means the action is
...

    ## Pagination

...

...

...

...

...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    PageToken:
      name: pageToken
      in: query
      schema: { type: string }
...
    TaskId:
      name: task
      in: path
      required: true
      description: The task_state ledger's task name (an AgentConfig name).
      schema: { type: string }
    CandidateKey:
      name: key
      in: path
      required: true
      description: The task_state ledger's candidate_key.
      schema: { type: string }
...
  schemas:
    Problem:
      type: object
      description: An RFC 9457 problem document, served as application/problem+json.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
      required: [title, status]
      example:
        title: run not found
        status: 404
...
    TaskStateItem:
      type: object
      description: |
        One row in the task_state ledger: the durable cross-run record of which
        candidates a task has handled and their outcome. task and candidateKey
        are opaque strings the deployment defines; outcome is opaque JSON the
        engine never interprets.
      properties:
        task: { type: string, description: "The AgentConfig (task) name." }
        candidateKey: { type: string, description: "Deployment-defined dedup identity for a work item (e.g. an external record id)." }
        status: { type: string, description: "One of: in_progress, done, skipped, escalated, needs_attention." }
        outcome:
          description: "Arbitrary structured result recorded by the supervisor, passed through untouched. Any valid JSON value: an object, array, string, number, boolean, or null."
          x-go-type: json.RawMessage
        caseRunId: { type: string, description: "Temporal run id of the Case workflow that claimed this candidate." }
        attempts: { type: integer, description: "Case start count; guards infinite re-dispatch of failing candidates." }
        createTime: { type: string, format: date-time }
        updateTime: { type: string, format: date-time }
      required: [task, candidateKey, status]
      example:
        task: "example-task"
        candidateKey: "record-1"
        status: done
        outcome: { result: "ok" }
        caseRunId: "b3f1c2a4-5d6e-4f80-9a1b-2c3d4e5f6071"
        attempts: 1
        createTime: "2026-07-08T12:00:00Z"
        updateTime: "2026-07-08T12:05:00Z"
    TaskStateList:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/TaskStateItem' } }
        nextPageToken: { type: string }
      required: [items]
...
security:
  - bearerAuth: []
paths:
...
  /agent-tasks/{task}/state:
    get:
      operationId: listTaskState
      description: A page of the task_state ledger for the given task, newest updated first.
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageToken'
        - { name: status, in: query, description: "Filter by status: in_progress, done, skipped, escalated, or needs_attention.", schema: { type: string } }
      responses:
        "200":
          description: a page of task_state ledger rows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskStateList' }
  /agent-tasks/{task}/state/{key}/requeue:
    post:
      operationId: requeueTaskState
      description: |
        Clears the ledger row for this candidate so the next monitor pass
        treats it as new and re-dispatches it. The explicit human action for
        returning a needs_attention (or done/skipped, for a forced redo)
        candidate to the queue.
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/CandidateKey'
      responses:
        "204":
          description: the candidate was requeued
        "404":
          description: no ledger row for this task and key
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...